// src/common/enums/notification.enum.ts
export enum NotificationType {
  CAR = 'car',
  CARSMEUP_CERTIFIED = 'carsmeup_certified',
  HOT_DEAL = 'hot_deal',
  BUMP = 'bump',
  ROADSIDE_ASSIST = 'roadside_assist',
  B_QUIK_BENZINE = 'b_quik_benzine',
  B_QUIK_DIESEL = 'b_quik_diesel',
}

export enum NotificationSubType {
  PUBLISHED = 'published',
  NOT_APPROVED = 'not_approved',
  EXPIRED = 'expired',
  DELETED = 'deleted',
  ACTION_REQUIRED = 'action_required',
  NEED_ACTION = 'need_action',
  PURCHASED = 'purchased',
}

export enum StaffNotificationType {
  DEALER = 'dealer',
}

export enum StaffNotificationSubType {
  REQUESTED = 'requested',
}